+++ NEWS.md
@@ -2,6 +2,10 @@
 
 ## Unversioned
 
+### Bugfixes
+
+* Fix a bug in which field names for the capacities of a `Storage` node resulted in error message for the investment data checks.
+
 ### Minor updates
 
 * Added checks
@@ -10,7 +14,7 @@
 
 ## Version 0.9.3 (2025-10-23)
 
-## Bugfixes
+### Bugfixes
 
 * Fix a bug in which field names for the capacities of a `Storage` node resulted in unconstrained capacities.